# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LUMINA social network API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts/feed:
    get:
      operationId: getFeed
      tags: [posts]
      summary: Get paginated feed for current user
      security:
        - bearerAuth: []
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Feed posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeedPost"
  /posts/explore:
    get:
      operationId: getExplorePosts
      tags: [posts]
      summary: Get trending posts
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: tag
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Trending posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeedPost"
  /posts/saved:
    get:
      operationId: getSavedPosts
      tags: [posts]
      summary: Get current user's saved posts
      security:
        - bearerAuth: []
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Saved posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeedPost"
...
  /posts:
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedPost"
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a single post with comments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /posts/{id}/like:
    post:
      operationId: togglePostLike
      tags: [posts]
      summary: Toggle like/reaction on a post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionInput"
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"
  /posts/{id}/save:
    post:
      operationId: togglePostSave
      tags: [posts]
      summary: Toggle save on a post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Save toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveResult"
  /posts/{id}/comments:
    get:
      operationId: getPostComments
      tags: [posts]
      summary: Get comments for a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"

    post:
      operationId: createComment
      tags: [posts]
      summary: Add a comment to a post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /posts/comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [posts]
      summary: Delete a comment
      security:
        - bearerAuth: []
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /posts/comments/{commentId}/like:
    post:
      operationId: toggleCommentLike
      tags: [posts]
      summary: Toggle like on a comment
      security:
        - bearerAuth: []
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"

  # ── USERS ───────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    UserSummary:
      type: object
      required: [id, username, displayName, avatarUrl, verified]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
        verified:
          type: boolean
        bio:
          type: string
        isFollowing:
          type: boolean
        followerCount:
          type: integer
...
    PostInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
        imageUrl:
          type: ["string", "null"]
        imageUrl2:
          type: ["string", "null"]
          description: Second image for Moment posts
        postType:
          type: string
          enum: [post, moment]
          default: post
        scheduledAt:
          type: ["string", "null"]
          format: date-time
    FeedPost:
      type: object
      required: [id, content, imageUrl, imageUrl2, postType, views, createdAt, author, likeCount, commentCount, liked, saved, topReactions, previewComments]
      properties:
        id:
          type: integer
        content:
          type: string
        imageUrl:
          type: string
        imageUrl2:
          type: string
        postType:
          type: string
        views:
          type: integer
        scheduledAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        author:
          $ref: "#/components/schemas/UserSummary"
        likeCount:
          type: integer
        commentCount:
          type: integer
        liked:
          type: boolean
        saved:
          type: boolean
        myReaction:
          type: ["string", "null"]
        topReactions:
          type: array
          items:
            $ref: "#/components/schemas/ReactionCount"
        previewComments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
    PostDetail:
      type: object
      required: [id, content, imageUrl, imageUrl2, postType, views, createdAt, author, likeCount, commentCount, liked, saved, topReactions, comments]
      properties:
        id:
          type: integer
        content:
          type: string
        imageUrl:
          type: string
        imageUrl2:
          type: string
        postType:
          type: string
        views:
          type: integer
        createdAt:
          type: string
          format: date-time
        author:
          $ref: "#/components/schemas/UserSummary"
        likeCount:
          type: integer
        commentCount:
          type: integer
        liked:
          type: boolean
        saved:
          type: boolean
        myReaction:
          type: ["string", "null"]
        topReactions:
          type: array
          items:
            $ref: "#/components/schemas/ReactionCount"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
    ReactionCount:
      type: object
      required: [reaction, count]
      properties:
        reaction:
          type: string
        count:
          type: integer
    LikeResult:
      type: object
      required: [liked, count]
      properties:
        liked:
          type: boolean
        count:
          type: integer
        reaction:
          type: ["string", "null"]
    SaveResult:
      type: object
      required: [saved]
      properties:
        saved:
          type: boolean

    # ── Comment schemas ──────────────────────────────────────────────────────
    CommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
        parentCommentId:
          type: ["integer", "null"]
    Comment:
      type: object
      required: [id, postId, content, createdAt, author, likeCount, liked]
      properties:
        id:
          type: integer
        postId:
          type: integer
        content:
          type: string
        parentCommentId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        author:
          $ref: "#/components/schemas/UserSummary"
        likeCount:
          type: integer
        liked:
          type: boolean

    # ── Follow / Reaction schemas ────────────────────────────────────────────
...
    ReactionInput:
      type: object
      properties:
        reaction:
          type: ["string", "null"]
          description: "Emoji: ❤️ 🔥 😂 😮 😢 👏 or null for simple like"
...