# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LUMINA social network API
 * OpenAPI spec version: 0.1.0
 */
import type { UserSummary } from './userSummary';

export interface Comment {
  id: number;
  postId: number;
  content: string;
  /** @nullable */
  parentCommentId?: number | null;
  createdAt: Date;
  author: UserSummary;
  likeCount: number;
  liked: boolean;
}