# lib/api-zod/src/generated/types/feedPost.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LUMINA social network API
 * OpenAPI spec version: 0.1.0
 */
import type { Comment } from './comment';
import type { ReactionCount } from './reactionCount';
import type { UserSummary } from './userSummary';

export interface FeedPost {
  id: number;
  content: string;
  imageUrl: string;
  imageUrl2: string;
  postType: string;
  views: number;
  /** @nullable */
  scheduledAt?: Date | null;
  createdAt: Date;
  author: UserSummary;
  likeCount: number;
  commentCount: number;
  liked: boolean;
  saved: boolean;
  /** @nullable */
  myReaction?: string | null;
  topReactions: ReactionCount[];
  previewComments: Comment[];
}